+++ test/src/typeset_test.dart
@@ -1,150 +1,46 @@
 // ignore_for_file: prefer_const_constructors
-
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:typeset/typeset.dart';
+import 'package:typeset/src/typeset_parser.dart';
+
+import 'typeset_widget.dart';
 
 //example widget for testing
-class _TypeSetTest extends StatelessWidget {
-  const _TypeSetTest({
-    super.key,
-    this.title,
-    this.style,
-    this.titleForExt,
-  });
-
-  final String? title;
-  final TextStyle? style;
-  final String? titleForExt;
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        body: Center(
-          child: Column(
-            children: [
-              if (title != null)
-                TypeSet(
-                  inputText: title!,
-                  style: style,
-                ),
-              if (titleForExt != null)
-                titleForExt!.typeset(
-                  style: style,
-                ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   group(
     'Tests for TypeSet Widget',
     () {
       testWidgets(
-        'TypeSet widget displays normal text correctly',
+        'TypeSet widget displays bold, italic, and strikethrough text',
         (WidgetTester tester) async {
           await tester.pumpWidget(
-            _TypeSetTest(title: 'Hello World'),
-          );
-
-          expect(
-            find.text(
-              'Hello World',
-              findRichText: true,
-            ),
-            findsOneWidget,
-          );
-        },
-      );
-
-      testWidgets(
-        'TypeSet widget displays bold text correctly',
-        (WidgetTester tester) async {
-          await tester.pumpWidget(
-            _TypeSetTest(
-              title: 'Hello, *World!*',
-              key: Key('boldText'),
-            ),
-          );
-          final boldText = find.byKey(
-            Key('boldText'),
-          );
-          expect(
-            boldText,
-            findsOneWidget,
-          );
-        },
-      );
-
-      testWidgets(
-        'TypeSet widget displays italic text correctly',
-        (WidgetTester tester) async {
-          await tester.pumpWidget(
-            _TypeSetTest(
-              title: 'Hello, _World_!',
-              key: Key('italicText'),
-            ),
-          );
-          final italicText = find.byKey(
-            Key('italicText'),
-          );
-          expect(italicText, findsOneWidget);
-        },
-      );
-
-      testWidgets(
-        'TypeSet widget displays underline text correctly',
-        (WidgetTester tester) async {
-          await tester.pumpWidget(
-            _TypeSetTest(
-              title: 'Hello, ~World~!',
-              key: Key('underlineText'),
-            ),
-          );
-          final underlineText = find.byKey(
-            Key('underlineText'),
-          );
-          expect(underlineText, findsOneWidget);
-        },
-      );
-
-      testWidgets(
-        'TypeSet widget displays underline text correctly',
-        (WidgetTester tester) async {
-          await tester.pumpWidget(
-            _TypeSetTest(
-              title: 'Hello, -World-!',
-              key: Key('underlineText'),
-            ),
-          );
-          final underlineText = find.byKey(
-            Key('underlineText'),
-          );
-          expect(underlineText, findsOneWidget);
-        },
-      );
-
-      testWidgets(
-        'TypeSet widget displays bold, italic, and underline text correctly',
-        (WidgetTester tester) async {
-          await tester.pumpWidget(
-            _TypeSetTest(
-              title: 'Hello, *World* _World_ ~World~!',
-              key: Key('boldItalicUnderlineText'),
-              style: TextStyle(color: Colors.black, fontSize: 20),
+            TypeSetTest(
+              title:
+                  'Hello, *World* _World_ ~World~ //hello// [hello](https://google.com)',
+              key: Key(
+                'typeset_widget_test',
+              ),
+              style: TextStyle(
+                color: Colors.black,
+                fontSize: 20,
+              ),
             ),
           );
 
           final boldItalicUnderlineText = find.byKey(
-            Key('boldItalicUnderlineText'),
+            Key(
+              'typeset_widget_test',
+            ),
           );
 
-          expect(boldItalicUnderlineText, findsOneWidget);
+          expect(
+            boldItalicUnderlineText,
+            findsOneWidget,
+          );
         },
       );
     },
@@ -155,7 +51,7 @@ void main() {
       'TypeSet widget displays through extension',
       (WidgetTester tester) async {
         await tester.pumpWidget(
-          _TypeSetTest(
+          TypeSetTest(
             titleForExt: 'Hello World',
             key: Key(
               'extensionTest',
@@ -174,5 +70,132 @@ void main() {
         );
       },
     );
+  });
+
+  group('TypesetParser', () {
+    test('parses bold text', () {
+      const inputText = 'This *word* is bold';
+      final expectedSpans = [
+        TextSpan(
+          text: 'This ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'word ',
+          style: TextStyle(
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        TextSpan(
+          text: 'is ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'bold',
+          style: TextStyle(),
+        )
+      ];
+      expect(
+        TypesetParser.parseText(inputText),
+        equals(expectedSpans),
+      );
+    });
+
+    test('parses italic text', () {
+      const inputText = 'This _word_ is italic';
+      final expectedSpans = [
+        TextSpan(
+          text: 'This ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'word ',
+          style: TextStyle(
+            fontStyle: FontStyle.italic,
+          ),
+        ),
+        TextSpan(
+          text: 'is ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'italic',
+          style: TextStyle(),
+        )
+      ];
+      expect(
+        TypesetParser.parseText(inputText),
+        equals(expectedSpans),
+      );
+    });
+
+    test('parses underline text', () {
+      const inputText = 'This //word// is underlined';
+      final expectedSpans = [
+        TextSpan(
+          text: 'This ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'word ',
+          style: TextStyle(
+            decoration: TextDecoration.underline,
+          ),
+        ),
+        TextSpan(
+          text: 'is ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'underlined',
+          style: TextStyle(),
+        )
+      ];
+      expect(
+        TypesetParser.parseText(inputText),
+        equals(expectedSpans),
+      );
+    });
+
+    test('parses strikethrough text', () {
+      const inputText = 'This ~word~ is strikethrough';
+      final expectedSpans = [
+        TextSpan(
+          text: 'This ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'word ',
+          style: TextStyle(
+            decoration: TextDecoration.lineThrough,
+          ),
+        ),
+        TextSpan(
+          text: 'is ',
+          style: TextStyle(),
+        ),
+        TextSpan(
+          text: 'strikethrough',
+          style: TextStyle(),
+        )
+      ];
+      expect(
+        TypesetParser.parseText(inputText),
+        equals(expectedSpans),
+      );
+    });
+
+    test('parses link correctly', () {
+      const inputText = '[Example|http://example.com]';
+
+      final result = TypesetParser.parseText(inputText);
+
+      expect(result.length, 1);
+      final linkSpan = result[0];
+      expect(linkSpan.text, 'Example');
+      expect(linkSpan.style?.color, Colors.blue);
+      expect(linkSpan.style?.decoration, TextDecoration.underline);
+      expect(linkSpan.recognizer, isA<TapGestureRecognizer>());
+    });
   });
 }